requirement.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
ciscoisesdk==2.2.3
click==8.1.8
colorama==0.4.6
coverage==7.6.10
cryptography==44.0.0
dict2xml==1.7.6
dill==0.3.9
f5-icontrol-rest==1.3.13
fastjsonschema==2.21.1
Flask==3.1.0
frozenlist==1.5.0
future==1.0.0
genie==24.11
genie.libs.clean==24.11
genie.libs.conf==24.11
genie.libs.filetransferutils==24.11
genie.libs.health==24.11
genie.libs.ops==24.11
genie.libs.parser==24.11
genie.libs.sdk==24.11
grpcio==1.69.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonpickle==4.0.1
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
meraki==1.53.0
multidict==6.1.0
mypy-extensions==1.0.0
ncclient==0.6.16
netaddr==0.10.1
netmiko==4.5.0
nornir==3.4.1
nornir-netmiko==1.0.1
nornir-utils==0.2.0
ntc_templates==7.5.0
packaging==24.2
paramiko==3.5.0
pluggy==1.5.0
ply==3.11
prettytable==3.12.0
propcache==0.2.1
protobuf==5.29.2
pyasn1==0.4.8
pyasynchat==1.0.4
pyasyncore==1.0.4
pycparser==2.22
pyftpdlib==2.0.1
Pygments==2.19.0
PyJWT==1.7.1
PyNaCl==1.5.0
pyserial==3.5
pysmi==1.5.9
pysnmp==6.1.4
pysnmpcrypto==0.0.4
pytest==8.3.4
pytest-cov==4.1.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rest.connector==24.11
rich==13.9.4
ruamel.yaml==0.18.9
ruamel.yaml.clib==0.2.12
scp==0.15.0
setuptools==75.7.0
six==1.17.0
textfsm==1.1.3
tftpy==0.8.0
tqdm==4.67.1
unicon==24.11
unicon.plugins==24.11
urllib3==2.3.0
wcwidth==0.2.13
webexteamssdk==1.7
Werkzeug==3.1.3
wheel==0.45.1
xmltodict==0.12.0
yang.connector==24.11
yarl==1.18.3
